+++ pharma/src/main/java/com/example/pharma/dto/review/ReviewDto.java
@@ -3,6 +3,6 @@ package com.example.pharma.dto.review;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 public record ReviewDto(@JsonProperty("name") String name,
-                        @JsonProperty("rating") Integer rating,
+                        @JsonProperty("rating") Long rating,
                         @JsonProperty("comment") String comment) {
 }
