+++ pharma/src/main/java/com/example/pharma/repository/Core/OwnerProfileRepository.java
@@ -3,4 +3,4 @@ package com.example.pharma.repository.Core;
 import com.example.pharma.model.entity.core.OwnerProfile;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OwnerProfileRepository extends JpaRepository<OwnerProfile, Integer> {}
+public interface OwnerProfileRepository extends JpaRepository<OwnerProfile, Long> {}
