+++ pharma/src/main/java/com/example/pharma/repository/Order/OrderRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface OrderRepository extends JpaRepository<Order, Integer> {
-    List<Order> findByCustomer_UserId(Integer userId);
-    List<Order> findByPharmacy_PharmacyId(Integer pharmacyId);
+public interface OrderRepository extends JpaRepository<Order, Long> {
+    List<Order> findByCustomer_UserId(Long userId);
+    List<Order> findByPharmacy_PharmacyId(Long pharmacyId);
 }
