+++ pharma/src/main/java/com/example/pharma/repository/Review/SellerRatingRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface SellerRatingRepository extends JpaRepository<SellerRating, Integer> {
-    List<SellerRating> findBySeller_UserId(Integer userId);
-    List<SellerRating> findByBuyer_UserId(Integer userId);
+public interface SellerRatingRepository extends JpaRepository<SellerRating, Long> {
+    List<SellerRating> findBySeller_UserId(Long userId);
+    List<SellerRating> findByBuyer_UserId(Long userId);
 }
